+++ test/Sentry.Azure.Functions.Worker.Tests/ApiApprovalTests.verify.cs
@@ -1,4 +1,4 @@
-namespace Sentry.AzureFunctions.Worker.Tests;
+namespace Sentry.Azure.Functions.Worker.Tests;
 
 [UsesVerify]
 public class ApiApprovalTests